fix(email db): gh pages is static lol
