--- deno.json
@@ -1,5 +1,0 @@
-{
-  "imports": {
-    "@/": "./src/"
-  }
-}
